Lower OpenGL version to 3.3
We don't need the newer fancy calls, and I want Papertoy to work on
more systems.

This doesn't do a full cleaning of everything that depends on 4.5 but
it works as long as you don't touch anything that would reference a
GL>3.3 call (enums are not lazy and therefore are correct).